requirements.txt
aptos-sdk
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
flask-cors
pythclient  
Flask-APScheduler
requests